remove a lot of JavaScript
First, lots of Flickr images had this embedding-related script that
would draw a Flickr logo and give options that I did not need anybody to
have.  Also, I removed the height and width attributes, because the
images are already the right size!

I also removed Twitter JavaScript (ugh) because it stopped working
anyway and also Twitter.